move files into contextual folders and change imports to use folder level index [elifesciences/enhanced-preprints-issues#1107]
