+++ src/components/atoms/article-content/article-content.tsx
@@ -1,5 +1,5 @@
 import './article-content.scss';
-import { JSXContent } from '../../../utils/content-to-jsx';
+import { JSXContent } from '../../../utils/content';
 
 export const ArticleContent = ({ content }: { content: JSXContent }) => (
   <article className="article-body">{content}</article>
