+++ src/utils/content/content-to-html.tsx
@@ -1,5 +1,5 @@
 import ReactDOMServer from 'react-dom/server';
-import { Content } from '../types';
+import { Content } from '../../types';
 import { contentToJsx } from './content-to-jsx';
 
 export const contentToHtml = (content: Content): string => {
