+++ src/utils/content/content-to-text.ts
@@ -1,4 +1,4 @@
-import { Content } from '../types';
+import { Content } from '../../types';
 
 export const contentToText = (content: Content): string => {
   if (typeof content === 'undefined') {
