+++ src/utils/data-fetch/fetch-data.ts
@@ -1,7 +1,7 @@
-import { config } from '../config';
+import { config } from '../../config';
 import { jsonFetch, jsonFetchOrNull } from './json-fetch';
-import { ArticleSummary, EnhancedArticleWithVersions } from '../types';
-import { EnhancedArticleNoContent } from '../types/reviewed-preprint-snippet';
+import { ArticleSummary, EnhancedArticleWithVersions } from '../../types';
+import { EnhancedArticleNoContent } from '../../types/reviewed-preprint-snippet';
 
 export const fetchVersion = (id: string, preview: boolean = false) => jsonFetchOrNull<EnhancedArticleWithVersions>(`${config.apiServer}/api/preprints/${id}${preview ? '?previews=true' : ''}`);
 export const fetchVersions = () => jsonFetch<{ items: ArticleSummary[], total: number }>(`${config.apiServer}/api/preprints`);
