+++ src/utils/formatters/format-author-name.ts
@@ -1,3 +1,3 @@
-import { Author } from '../types';
+import { Author } from '../../types';
 
 export const formatAuthorName = ({ givenNames, familyNames, honorificSuffix }: Author) => `${(givenNames ?? []).join(' ')} ${(familyNames ?? []).join(' ')}${honorificSuffix ? ` ${honorificSuffix}` : ''}`.trim();
