+++ src/utils/generators/generate-image-url.test.ts
